refactor: 레시피 목록 UiModel Id 타입 Long 으로 변경

+++ app/src/main/java/com/hyeok/recipebook/presentation/recipe/list/RecipeScreen.kt
@@ -60,7 +60,7 @@ fun RecipeScreen(
     recipesUiState: RecipesUiState,
     searchQueryState: TextFieldState,
     onAddRecipe: () -> Unit = {},
-    onClickRecipe: (Int) -> Unit = {},
+    onClickRecipe: (Long) -> Unit = {},
     onSelectFilter: (RecipeSearchFilter) -> Unit = {},
     onSearchRecipes: (String) -> Unit = {},
 ) {
